VideoCore: Move software rasterizer files to sub-directory
