+++ src/video_core/CMakeLists.txt
@@ -1,10 +1,8 @@
 set(SRCS
-            clipper.cpp
             command_processor.cpp
             debug_utils/debug_utils.cpp
             pica.cpp
             primitive_assembly.cpp
-            rasterizer.cpp
             regs.cpp
             renderer_base.cpp
             renderer_opengl/gl_rasterizer.cpp
@@ -15,7 +13,9 @@ set(SRCS
             renderer_opengl/renderer_opengl.cpp
             shader/shader.cpp
             shader/shader_interpreter.cpp
-            swrasterizer.cpp
+            swrasterizer/clipper.cpp
+            swrasterizer/rasterizer.cpp
+            swrasterizer/swrasterizer.cpp
             texture/etc1.cpp
             texture/texture_decode.cpp
             vertex_loader.cpp
@@ -23,7 +23,6 @@ set(SRCS
             )
 
 set(HEADERS
-            clipper.h
             command_processor.h
             debug_utils/debug_utils.h
             gpu_debugger.h
@@ -31,7 +30,6 @@ set(HEADERS
             pica_state.h
             pica_types.h
             primitive_assembly.h
-            rasterizer.h
             rasterizer_interface.h
             regs.h
             regs_framebuffer.h
@@ -52,7 +50,9 @@ set(HEADERS
             shader/debug_data.h
             shader/shader.h
             shader/shader_interpreter.h
-            swrasterizer.h
+            swrasterizer/clipper.h
+            swrasterizer/rasterizer.h
+            swrasterizer/swrasterizer.h
             texture/etc1.h
             texture/texture_decode.h
             utils.h
