+++ src/video_core/swrasterizer/clipper.h
