+++ src/video_core/swrasterizer/rasterizer.h
